Clean up code

+++ src/main/java/de/nihas101/image_to_pdf_converter/gui/controller/OptionsMenuController.java
@@ -139,7 +139,7 @@ public class OptionsMenuController {
     }
 
     public ImageToPdfOptions getImageToPdfOptions() {
-        imageToPdfOptions.setMaximalSearchDepth((Integer) maximalDepthSearchSpinner.getValue());
+        imageToPdfOptions.setMaximalSearchDepth(maximalDepthSearchSpinner.getValue());
         return imageToPdfOptions;
     }
